refactor(FT) move Attendee functionality to dedicated controller
